Added the nested query option
Added the nested query option, to adjust the "problem" of limit and offset in relation-dependent queries.

A nested subquery is used to return data that will be used in the main query as a condition to further restrict the data to be retrieved.
Source: https://www.postgresql.org/docs/8.4/functions-subquery.html